docs: add examples for components
